move all lobby files in dirs with function name like join (for joinLobby)
