+++ app/routes/games/lobby/join/joinLobby.ts
@@ -1,11 +1,11 @@
 import { FastifyPluginAsync } from "fastify";
-import joinLobbySchema from "../../../schemas/games/lobby/joinLobbySchema";
-import joinLobbyHandler from "../../../services/games/lobby/joinLobbyHandler";
+import joinLobbySchema from "../../../../schemas/games/lobby/joinLobbySchema";
+import joinLobbyHandler from "../../../../services/games/lobby/join/joinLobbyHandler";
 
 const joinLobby: FastifyPluginAsync = async (
   fastify
   ): Promise<void> => {
-  fastify.post("/join", {
+  fastify.post("/", {
     schema: joinLobbySchema,
     handler: joinLobbyHandler,
   });
