+++ app/routes/games/lobby/start/startLobby.ts
@@ -1,11 +1,11 @@
 import { FastifyPluginAsync } from "fastify";
-import startLobbySchema from "../../../schemas/games/lobby/startLobbySchema";
-import startLobbyHandler from "../../../services/games/lobby/startLobbyHandler";
+import startLobbySchema from "../../../../schemas/games/lobby/startLobbySchema";
+import startLobbyHandler from "../../../../services/games/lobby/start/startLobbyHandler";
 
 const startLobby: FastifyPluginAsync = async (
   fastify
   ): Promise<void> => {
-  fastify.post("/start", {
+  fastify.post("/", {
     schema: startLobbySchema,
     handler: startLobbyHandler,
   });
